Move transactor to new module
