+++ core/src/main/scala/slickeffect/implicits.scala
